updated turn seq
updated so that each turn the travel toll is applied to the stats and the HUD is updated. Also added message if player is sick and the resulting loss of stats. Added a addiotnalInfo variable to concat to User message call so we can send additional info to the user after dice rolls.